moved handleLocal function to reducer
